Found suitable solution for hand IK and have working positional change
After many tries using different solutions, I have found one that seems to work as intended so far. For this one I am using a sphere collider coming from the shoulder and encompassing it to represent the zone that the hand IK can occur in this project. Once the sphere hits an object, I compare the hit location to several XYZ value limits to make sure it only works when the object and player are within a certain distance range. The hand is then moved to this position once it has passed all the "tests". I do this for each shoulder/hand. Will need to fix the hand and arm rotation, as these do not work as intended just yet.