Attempt to fix hand IK
From working on the hand IK, I have gotten it to a point where it somewhat works. The hand will only reach for the wall if the hit position is within the correct bounds and the blending in and out of IK is decent. The problem is that when not idle, the IK messes up, going back and forth which just doesnt look good. I have spent too much time on this so I will be leaving the hand IK for now. The blueprints need some cleanup but I will leave it in its current state for when I come back to it.