Stealth takedown implemented
When crouched and within range behind the target, the player can now perform a stealth takedown by pressing the f key. Will need to add prompts or someway to tell the player which key to press for a takedown and may also need visual queues such as the player or enemy glowing. Need to clean up the directories, replacing the old stuff with the new.