Fixed issue of enemy respawning with no path but not tagged as roaming
Enemies are either roaming or not roaming with non roaming enemies needing a set path to follow. Enemies by default are not roaming and are given a path when placed in the level however enemies respawning are given no path. To fix this I just check on spawn if their path array is empty and if it is, they become a roaming enemy.