Implemented player respawn
When the player is caught by an enemy, they will play a short animation and be transported back to the middle of the map to try again. There is still a vision cone animation bug where upon respawning, while the vision cone animation is playing, it will become jittery and after the vision cone animation, the player is unable to move and turning causes the vision cone to glitch which includes it shaking and parts of it being made black that shouldn't be.